Split out CreateJobOrder and fixed the bug where text input was causing lag
